change auth logic #44
