--- app/routes/__authorized/already_logged_in.tsx
@@ -1,9 +1,0 @@
-import { useEffect } from "react";
-
-import { message } from "antd";
-
-export default function Already() {
-  useEffect(() => {
-    message.info("You are already logged in.");
-  }, []);
-}
