+++ app/routes/__authorized/q/edit/$publicId.tsx
@@ -1,4 +1,4 @@
-import { Link, useCatch, useFetcher, useLoaderData } from "@remix-run/react";
+import { useFetcher, useLoaderData } from "@remix-run/react";
 import type {
   ActionArgs,
   LoaderArgs,
@@ -341,47 +341,3 @@ export const action = async ({ request, params }: ActionArgs) => {
     error: null,
   });
 };
-
-export function CatchBoundary() {
-  const caught = useCatch();
-
-  return (
-    <div className="min-h-full px-6 py-16 bg-white @sm:py-24 @md:grid @md:place-items-center @lg:px-8">
-      <div className="mx-auto max-w-max">
-        <main className="@sm:flex">
-          <p className="text-4xl font-bold tracking-tight text-green-600 @sm:text-5xl">
-            {caught.status}
-          </p>
-          <div className="@sm:ml-6">
-            <div className="@sm:border-l @sm:border-gray-200 @sm:pl-6">
-              <h1 className="text-4xl font-bold tracking-tight text-gray-900 @sm:text-5xl">
-                {caught.status === 404
-                  ? "페이지를 찾을 수 없습니다."
-                  : "오류가 발생했습니다."}
-              </h1>
-              {caught.status === 404 && (
-                <p className="mt-1 text-base text-gray-500">
-                  URL이 올바른지 확인하고 다시 시도해주세요.
-                </p>
-              )}
-            </div>
-            <div className="flex mt-10 space-x-3 @sm:border-l @sm:border-transparent @sm:pl-6">
-              <Link
-                to="/"
-                className="inline-flex items-center px-4 py-2 text-sm font-medium text-white bg-green-600 border border-transparent rounded-md shadow-sm hover:bg-green-700 focus:outline-none focus:ring-2 focus:ring-green-500 focus:ring-offset-2 "
-              >
-                홈으로 돌아가기
-              </Link>
-              <Link
-                to="/account"
-                className="inline-flex items-center px-4 py-2 text-sm font-medium text-green-700 bg-green-100 border border-transparent rounded-md hover:bg-green-200 focus:outline-none focus:ring-2 focus:ring-green-500 focus:ring-offset-2"
-              >
-                문의하기
-              </Link>
-            </div>
-          </div>
-        </main>
-      </div>
-    </div>
-  );
-}
